Graph Payment type Int to Float

+++ backend/app/graphql/types/payment_type.rb
@@ -4,7 +4,7 @@ module Types
     field :id, ID, null: false
     field :recipient, Types::UserType, null: false
     field :sender, Types::UserType, null: false
-    field :amount, Int, null: false
-    
+    field :amount, Float, null: false
+
   end
 end
